Precompute Hunspell dictionary explosion
