--- dict/da-DK/da-DK.dic.censored
@@ -1,18 +1,0 @@
-aftne/1,21,22,79,39
-fanme
-fucke/1,21,22,79,39
-fuckfinger/14,10,13,39
-fucking/39
-kvabso/44,10,39
-kvabsøer/74,10,39
-møgso/44,10,39,86
-møgsøer/74,10,39,86
-neger/14,10,13,39
-negere
-negeres
-negerne
-negernes
-pis/10,57,39
-pisse/1,21,22,79,39
-smatso/44,10,39
-smatsøer/74,10,39
