+++ dict/da-DK/da-DK.txt.censored
@@ -1,0 +1,82 @@
+aftn
+aftne
+aftnede
+aftnedes
+aftnen
+aftnende
+aftnendes
+aftnens
+aftner
+aftnes
+aftnet
+aftnets
+fanme
+fuck
+fucke
+fuckede
+fuckedes
+fucken
+fuckende
+fuckendes
+fuckens
+fucker
+fuckes
+fucket
+fuckets
+fuckfinger
+fuckfingeren
+fuckfingerens
+fuckfingers
+fuckfingre
+fuckfingrene
+fuckfingrenes
+fuckfingres
+fucking
+kvabso
+kvabsoen
+kvabsoens
+kvabsos
+kvabsøer
+kvabsøerne
+kvabsøernes
+kvabsøers
+møgso
+møgsoen
+møgsoens
+møgsos
+møgsøer
+møgsøerne
+møgsøernes
+møgsøers
+neger
+negere
+negeren
+negerens
+negeres
+negerne
+negernes
+negers
+negre
+negrene
+negrenes
+negres
+pis
+pisse
+pissede
+pissedes
+pissen
+pissende
+pissendes
+pissens
+pisser
+pisses
+pisset
+pissets
+smatso
+smatsoen
+smatsoens
+smatsos
+smatsøer
+smatsøerne
+smatsøernes
+smatsøers
