--- dict/en-GB/en-GB.dic.censored
@@ -1,19 +1,0 @@
-dago/S
-dagoes
-lesbigay/S
-lesbo/S
-lez
-lezzes
-lezzie/S
-nigger/S
-poof/S
-poofter/S
-spic/S
-spick/S
-tarbrush/S
-tranny/S
-whore/S,D,G
-whoredom
-whoreson/S
-whorish/Y
-wog/S
