+++ dict/en-GB/en-GB.txt.censored
@@ -1,0 +1,36 @@
+dago
+dagoes
+dagos
+lesbigay
+lesbigays
+lesbo
+lesbos
+lez
+lezzes
+lezzie
+lezzies
+nigger
+niggers
+poof
+poofs
+poofter
+poofters
+spic
+spick
+spicks
+spics
+tarbrush
+tarbrushes
+trannies
+tranny
+whore
+whored
+whoredom
+whores
+whoreson
+whoresons
+whoring
+whorish
+whorishly
+wog
+wogs
