+++ game_gen/CMakeLists.txt
@@ -1,4 +1,6 @@
 cmake_minimum_required(VERSION 3.10)
 
+project (game_generator)
+
 add_subdirectory (external/anatree anatree)
 add_subdirectory (src)
